robot/audio/pyproject.toml
[project]
dependencies = [
  "edge-tts>=7.2.3",
  "numpy>=2.0.0",
  "pvporcupine>=3.0.5",
  "pyaudio>=0.2.14",
  "pygame>=2.6.1",
  "pytest>=9.0.1",
  "torch>=2.9.0",
  "transformers>=4.40.0",
  "eclipse-zenoh>=1.6.2",
  "protobuf>=5.29.3",
  "grpcio-tools>=1.70.1",
]
description = "Audio component for wake word detection and TTS"
name = "audio"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"
